Implemented requested changes.
