+++ src/MvvmNano.Forms/MvvmNano.Forms.csproj
@@ -38,7 +38,6 @@
     <Compile Include="MvvmNanoFormsPresenter.cs" />
     <Compile Include="MvvmNanoMasterDetailApplication.cs" />
     <Compile Include="MvvmNanoMasterDetailPage.cs" />
-    <Compile Include="NanoMasterDetailPage.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="MvvmNanoContentPage.cs" />
     <Compile Include="MvvmNanoNavigationPage.cs" />
